Add 'commit'
This allows a commit to be created. The commit is always done on the
root of the repository, and it is currently not possible to ignore any
files. Author and committer (always the same) are taken from user.email
and user.name in .git/config.

Commits are always done on the head of master. No other branch is
supported yet.